Renames reference to AddErrorToPageTitle hook
