+++ src/components/forms/TextInput/index.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import GDSTextInput from '../../BaseComponents/TextInput/TextInput';
-import usePrefixPageTitle from '../../../helpers/hooks/usePrefixPageTitle';
+import useAddErrorToPageTitle from '../../../helpers/hooks/useAddErrorToPageTitle';
 
 declare const PCore;
 
@@ -21,7 +21,7 @@ export default function TextInput(props) {
 
   const [isOnlyQuestion, setIsOnlyQuestion] = useState(PCore.getFormUtils().getEditableFields("root/primary_1/workarea_1").length === 1);
   const [displayLabel, setDisplayLabel] = useState(label);
-  usePrefixPageTitle(validatemessage);
+  useAddErrorToPageTitle(validatemessage);
 
   /* let testProp = {};
   testProp = {
