+++ src/helpers/hooks/useAddErrorToPageTitle.tsx
